migrate the config.go file from agent into an HCL package, including the ParseHCL fn and tests
