+++ agent/agent_test.go
@@ -233,131 +233,6 @@ func TestSetup(t *testing.T) {
 	}
 }
 
-func TestParseHCL(t *testing.T) {
-	testCases := []struct {
-		name   string
-		path   string
-		expect Config
-	}{
-		{
-			name:   "Empty config is valid",
-			path:   "../tests/resources/config/empty.hcl",
-			expect: Config{},
-		},
-		{
-			name: "Host with no attributes is valid",
-			path: "../tests/resources/config/host_no_ops.hcl",
-			expect: Config{
-				Host: &HostConfig{},
-			},
-		},
-		{
-			name: "Host with one of each op is valid",
-			path: "../tests/resources/config/host_each_op.hcl",
-			expect: Config{
-				Host: &HostConfig{
-					Commands: []CommandConfig{
-						{Run: "testing", Format: "string"},
-					},
-					Shells: []ShellConfig{
-						{Run: "testing"},
-					},
-					GETs: []GETConfig{
-						{Path: "/v1/api/lol"},
-					},
-					Copies: []CopyConfig{
-						{Path: "./*", Since: "10h"},
-					},
-				},
-			},
-		},
-		{
-			name: "Host with multiple of a op type is valid",
-			path: "../tests/resources/config/multi_ops.hcl",
-			expect: Config{
-				Host: &HostConfig{
-					Commands: []CommandConfig{
-						{
-							Run:    "testing",
-							Format: "string",
-						},
-						{
-							Run:    "another one",
-							Format: "string",
-						},
-						{
-							Run:    "do a thing",
-							Format: "json",
-						},
-					},
-				},
-			},
-		},
-		{
-			name: "Config with a host and one product with everything is valid",
-			path: "../tests/resources/config/config.hcl",
-			expect: Config{
-				Host: &HostConfig{
-					Commands: []CommandConfig{
-						{Run: "ps aux", Format: "string"},
-					},
-				},
-				Products: []*ProductConfig{
-					{
-						Name: "consul",
-						Commands: []CommandConfig{
-							{Run: "consul version", Format: "json"},
-							{Run: "consul operator raft list-peers", Format: "json"},
-						},
-						Shells: []ShellConfig{
-							{Run: "consul members | grep ."},
-						},
-						GETs: []GETConfig{
-							{Path: "/v1/api/metrics?format=prometheus"},
-						},
-						Copies: []CopyConfig{
-							{Path: "/another/test/log", Since: "240h"},
-						},
-						Excludes: []string{"consul some-awfully-long-command"},
-						Selects: []string{
-							"consul just this",
-							"consul and this",
-						},
-					},
-				},
-			},
-		},
-		{
-			name: "Config with multiple products is valid",
-			path: "../tests/resources/config/multi_product.hcl",
-			expect: Config{
-				Products: []*ProductConfig{
-					{
-						Name:     "consul",
-						Commands: []CommandConfig{{Run: "consul version", Format: "string"}},
-					},
-					{
-						Name:     "nomad",
-						Commands: []CommandConfig{{Run: "nomad version", Format: "string"}},
-					},
-					{
-						Name:     "vault",
-						Commands: []CommandConfig{{Run: "vault version", Format: "string"}},
-					},
-				},
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			res, err := ParseHCL(tc.path)
-			assert.NoError(t, err)
-			assert.Equal(t, tc.expect, res, tc.name)
-		})
-	}
-}
-
 func TestAgent_WriteSummary(t *testing.T) {
 	// NOTE: If you make changes to WriteSummary, you may break existing unit tests until the golden files are updated
 	// to reflect your changes. To update them, run `go test ./agent -update`, and then manually verify that the new
